Refactoring Fusion Executor, separating out compilation of a kernel (#3468)
Pull out kernel compilation from the `KernelExecutor`, separating out the two concepts as we move towards execution of a kernel through `HostIr`.

- Made `CompiledKernel` class to hold compilation information in `compiled_kernel.h/cpp`
- Moved code:
  - from `runtime/executor.h` to `runtime/compiled_kernel.h`
  - from `runtime/executor.cpp` to `runtime/compiled_kernel.cpp`
- from `runtime/executor_utils.cpp` to `runtime/compiled_kernel.cpp` (these are functions only used in `compiled_kernel`)
- from `sys/utils.cpp` to `runtime/compiled_kernel.cpp` (these are functions only used in `compiled_kernel`)
- Moved `executor::compileRTC` and `executor::runRTC` into its own class (`RtcKernel`). It shares compilation logic with `CompiledKernel` and is in `compiled_kernel.h/cpp`
