+++ csrc/runtime/executor_utils.h
@@ -29,21 +29,10 @@
 namespace nvfuser {
 namespace executor_utils {
 
-// Include all the functions we might need in generated code
-std::string kernelPreamble();
-
-//! Bind input values to runtime values
-NVF_API ExpressionEvaluator
-bindInputs(const KernelArgumentHolder& args, Fusion* fusion);
-
-NVF_API std::string disassembleBinary(
-    const std::vector<char>& cubin,
-    const std::string& nvdisasm_args);
-
-// I'm not happy with CompiledKernel being a struct exposing all the fields.
+// I'm not happy with CudaExecutable being a struct exposing all the fields.
 // This could be refactored.
-struct CompiledKernel : public NonCopyable {
-  NVF_API ~CompiledKernel();
+struct CudaExecutable : public NonCopyable {
+  NVF_API ~CudaExecutable();
 
   CUmodule module = nullptr;
   CUfunction function = nullptr;
@@ -58,20 +47,11 @@ struct CompiledKernel : public NonCopyable {
   int register_spills = -1;
 };
 
-// Returns executable function and the ptxas log from compilation
-std::unique_ptr<CompiledKernel> getCompiledKernel(
-    std::optional<std::reference_wrapper<const std::string>> kernel_code,
-    const std::string& code,
-    const std::string& func_name,
-    const std::string& id,
-    const CompileParams& compile_params = CompileParams(),
-    std::optional<int64_t> opt_block_size = std::nullopt);
+//! Bind input values to runtime values
+NVF_API ExpressionEvaluator
+bindInputs(const KernelArgumentHolder& args, Fusion* fusion);
 
-// Returns executable function using flatbuffer object
-std::unique_ptr<CompiledKernel> getCompiledKernel(
-    const serde::CudaKernel* buffer,
-    const CompileParams& compile_params);
-
+// Compile time cache for execution
 namespace caching {
 // TODO: Could consider putting some of
 //  the logic in the common space and re-use
@@ -252,13 +232,6 @@ void validateVectorizedTensors(
 void validateCircularBuffering(
     kir::Kernel* kernel,
     ExpressionEvaluator& expr_eval);
-
-//! Query the target GPU version number NVRTC compiles CUDA kernels for
-void queryTargetGPUVersion(
-    const cudaDeviceProp* const prop,
-    int64_t& major,
-    int64_t& minor,
-    bool& compile_to_sass);
 
 } // namespace executor_utils
 } // namespace nvfuser
