+++ csrc/runtime/fusion_kernel_runtime.h
@@ -107,9 +107,6 @@ class FusionKernelRuntime {
   }
 
   //! Internal knob for profiling shape inference
-  void disableLaunchParamCache();
-
-  //! Internal knob for profiling shape inference
   void disableKernelLaunch();
 
   //! Returns if this runtime is segmented
