+++ tests/cpp/test_circular_buffering.cpp
@@ -731,7 +731,8 @@ TEST_P(CircularBufferingTest, SmemBlockGemmCache) {
   //   insertion to ensure ordering of circular buffered tensor access.
   // The check below makes sure that the sync is inserted so that the
   //   test isn't running on a race condition.
-  NVF_CHECK(ke.kernel()->summary().war_hazard_syncs_count > 0);
+  NVF_CHECK(
+      ke.compiledKernel()->kernel()->summary().war_hazard_syncs_count > 0);
 }
 
 // Vectorized reset test for circular buffered registers
