+++ tests/cpp/test_gpu1.cpp
@@ -5816,7 +5816,8 @@ TEST_F(NVFuserTest, FusionSmem_CUDA) {
 
   testValidate(&fusion, cg_outputs, aten_inputs, __LINE__, __FILE__);
 
-  NVF_CHECK(ke.kernel()->summary().war_hazard_syncs_count == 0);
+  NVF_CHECK(
+      ke.compiledKernel()->kernel()->summary().war_hazard_syncs_count == 0);
 }
 
 TEST_F(NVFuserTest, FusionSmemReduce_CUDA) {
@@ -5864,7 +5865,8 @@ TEST_F(NVFuserTest, FusionSmemReduce_CUDA) {
 
   testValidate(
       &fusion, cg_outputs, {aten_input}, {aten_output}, __LINE__, __FILE__);
-  NVF_CHECK(ke.kernel()->summary().war_hazard_syncs_count == 0);
+  NVF_CHECK(
+      ke.compiledKernel()->kernel()->summary().war_hazard_syncs_count == 0);
 }
 
 TEST_F(NVFuserTest, FusionSmemBlockGemm_CUDA) {
@@ -5935,7 +5937,8 @@ TEST_F(NVFuserTest, FusionSmemBlockGemm_CUDA) {
   testValidate(
       &fusion, cg_outputs, aten_inputs, {aten_output}, __LINE__, __FILE__);
 
-  NVF_CHECK(ke.kernel()->summary().war_hazard_syncs_count == 0);
+  NVF_CHECK(
+      ke.compiledKernel()->kernel()->summary().war_hazard_syncs_count == 0);
 }
 
 TEST_F(NVFuserTest, FusionSmemBlockGemmCache_CUDA) {
@@ -6024,7 +6027,8 @@ TEST_F(NVFuserTest, FusionSmemBlockGemmCache_CUDA) {
   testValidate(
       &fusion, cg_outputs, aten_inputs, {aten_output}, __LINE__, __FILE__);
 
-  NVF_CHECK(ke.kernel()->summary().war_hazard_syncs_count == 0);
+  NVF_CHECK(
+      ke.compiledKernel()->kernel()->summary().war_hazard_syncs_count == 0);
 }
 
 TEST_F(NVFuserTest, FusionSmemDynamicPersistentSoftmax2D_CUDA) {
@@ -6385,9 +6389,9 @@ TEST_F(NVFuserTest, FusionMagicSchedulerLayerNormalization_CUDA) {
 
   // tv11 and tv17 should not be predicated. See issue #496
   ASSERT_FALSE(PredicatedChecker::isPredicated(
-      11, cg_results.kernel_executor->kernel()));
+      11, cg_results.kernel_executor->compiledKernel()->kernel()));
   ASSERT_FALSE(PredicatedChecker::isPredicated(
-      17, cg_results.kernel_executor->kernel()));
+      17, cg_results.kernel_executor->compiledKernel()->kernel()));
 }
 
 TEST_F(NVFuserTest, FusionMagicSchedulerRMSNormalization_CUDA) {
@@ -7217,7 +7221,8 @@ TEST_F(NVFuserTest, FusionSmemDynamicReductionSymbolic_CUDA) {
       __FILE__,
       "",
       lparams);
-  NVF_CHECK(ke.kernel()->summary().war_hazard_syncs_count == 0);
+  NVF_CHECK(
+      ke.compiledKernel()->kernel()->summary().war_hazard_syncs_count == 0);
 }
 
 TEST_F(NVFuserTest, FusionSmemDynamicReductionSymbolicArg_CUDA) {
@@ -7281,7 +7286,8 @@ TEST_F(NVFuserTest, FusionSmemDynamicReductionSymbolicArg_CUDA) {
       "",
       lparams);
 
-  NVF_CHECK(ke.kernel()->summary().war_hazard_syncs_count == 0);
+  NVF_CHECK(
+      ke.compiledKernel()->kernel()->summary().war_hazard_syncs_count == 0);
 }
 
 TEST_F(NVFuserTest, FusionSmemDynamicPwiseMulSymbolicArgWAR_CUDA) {
@@ -7338,7 +7344,8 @@ TEST_F(NVFuserTest, FusionSmemDynamicPwiseMulSymbolicArgWAR_CUDA) {
   testValidate(
       &fusion, cg_outputs, aten_inputs, __LINE__, __FILE__, "", lparams);
 
-  NVF_CHECK(ke.kernel()->summary().war_hazard_syncs_count == 1);
+  NVF_CHECK(
+      ke.compiledKernel()->kernel()->summary().war_hazard_syncs_count == 1);
 }
 
 TEST_F(NVFuserTest, FusionSmemDynamicTiledGemm_CUDA) {
@@ -7464,7 +7471,8 @@ TEST_F(NVFuserTest, FusionSmemDynamicTiledGemm_CUDA) {
   testValidate(
       &fusion, cg_outputs, aten_inputs, {aten_output}, __LINE__, __FILE__);
 
-  NVF_CHECK(ke.kernel()->summary().war_hazard_syncs_count == 1);
+  NVF_CHECK(
+      ke.compiledKernel()->kernel()->summary().war_hazard_syncs_count == 1);
 }
 
 } // namespace nvfuser
