+++ tests/cpp/test_gpu_outer_reduction.cpp
@@ -119,8 +119,10 @@ TEST_F(OuterReductionTest, GroupedGridWelfordOuterOpt) {
     ke.compile(&fusion, aten_inputs);
 
     NVF_CHECK(
-        ke.kernel()->summary().has_outer_grouped_grid_welford ==
-            params.should_use_opt,
+        ke.compiledKernel()
+                ->kernel()
+                ->summary()
+                .has_outer_grouped_grid_welford == params.should_use_opt,
         (params.should_use_opt ? "Failed to use the optimized implementation"
                                : "Should not use the optimized implementation"),
         ": ",
@@ -923,7 +925,7 @@ void grid_persistent_batchnorm_manual(
       true);
 
   testValidate(
-      ke.kernel(),
+      ke.compiledKernel()->kernel(),
       {cg_outputs.at(2)},
       aten_inputs,
       {at_output},
@@ -1251,7 +1253,7 @@ void grid_persistent_batchnorm_bwd_manual(
       {true, true, true});
 
   testValidate(
-      ke.kernel(),
+      ke.compiledKernel()->kernel(),
       cg_outputs,
       aten_inputs,
       {std::get<0>(at_output), std::get<1>(at_output), std::get<2>(at_output)},
@@ -2187,14 +2189,15 @@ TEST_F(OuterReductionTest, IterGroupedBlockReduction) {
 
   // lowering & check iteration grouped reductions
   NVF_CHECK(
-      ke.kernel()->summary().has_iter_grouped_reductions,
+      ke.compiledKernel()->kernel()->summary().has_iter_grouped_reductions,
       "There must be iter domain grouped reductions.");
   NVF_CHECK(
-      ke.kernel()->summary().num_grouped_iterations == vect_factor,
+      ke.compiledKernel()->kernel()->summary().num_grouped_iterations ==
+          vect_factor,
       "Expected ",
       vect_factor,
       " grouped iterations, found ",
-      ke.kernel()->summary().num_grouped_iterations);
+      ke.compiledKernel()->kernel()->summary().num_grouped_iterations);
 
   testValidate(
       &fusion,
