+++ tests/cpp/test_persistent_buffer.cpp
@@ -1312,8 +1312,8 @@ TEST_F(PersistentBufferTest, SmemPersistent2DReduction) {
   KernelExecutor ke;
   ke.compile(fusion.get(), aten_inputs);
   // Shared memory access should be vectorized.
-  // getBankConflictInfo(ke.kernel()) triggers error "std::get: wrong index for
-  // variant" when trying to evaluate index with:
+  // getBankConflictInfo(ke.compiledKernel()->kernel()) triggers error
+  // "std::get: wrong index for variant" when trying to evaluate index with:
   // `expr_eval.evaluate(ti->index()).as<int64_t>();`
   for (auto tv : fusion->allTvs()) {
     if (tv->getMemoryType() == MemoryType::Shared) {
