+++ tests/cpp/test_resize.cpp
@@ -1879,7 +1879,7 @@ TEST_F(ResizeTest, FusionSliceForNanoGPT1) {
   auto cg_outputs = executor_cache.runFusionWithInputs(aten_inputs);
 
   const auto* ke = onlyKernelExecutorInMostRecentRuntime(executor_cache);
-  auto kernel = ke->kernel();
+  auto kernel = ke->compiledKernel()->kernel();
   NVF_CHECK(
       !kernel->summary().has_cooperative_grid_reduction,
       "Grid sync should not be used as slicing input should avoid input caching");
@@ -1941,7 +1941,7 @@ TEST_F(ResizeTest, FusionSliceForNanoGPT2) {
   auto cg_outputs = executor_cache.runFusionWithInputs(aten_inputs);
 
   const auto* ke = onlyKernelExecutorInMostRecentRuntime(executor_cache);
-  auto kernel = ke->kernel();
+  auto kernel = ke->compiledKernel()->kernel();
 
   // Make sure the slices ops use the same producer
   TensorView* known_slice_producer = nullptr;
@@ -4148,8 +4148,11 @@ TEST_P(ResizeSchedulerTest, PropagateSliceToInputs) {
     const auto& heuristic_param =
         runtime->schedulerHeuristics()->heuristicsList().front();
     EXPECT_EQ(heuristic_param->scheduler_type, SchedulerType::Resize);
-    Fusion* scheduled_fusion =
-        runtime->executors().at(0)->as<KernelExecutor>()->kernel();
+    auto scheduled_fusion = runtime->executors()
+                                .at(0)
+                                ->as<KernelExecutor>()
+                                ->compiledKernel()
+                                ->kernel();
     checkLoopDomainEquivalence(
         scheduled_fusion->outputs().at(0)->as<TensorView>());
   }
@@ -4237,8 +4240,11 @@ TEST_P(ResizeSchedulerTest, PropagateSliceToInputsWithReshape1) {
     const auto& heuristic_param =
         runtime->schedulerHeuristics()->heuristicsList().front();
     EXPECT_EQ(heuristic_param->scheduler_type, SchedulerType::Resize);
-    Fusion* scheduled_fusion =
-        runtime->executors().at(0)->as<KernelExecutor>()->kernel();
+    auto scheduled_fusion = runtime->executors()
+                                .at(0)
+                                ->as<KernelExecutor>()
+                                ->compiledKernel()
+                                ->kernel();
     checkLoopDomainEquivalence(
         scheduled_fusion->outputs().at(0)->as<TensorView>());
   }
@@ -4322,8 +4328,11 @@ TEST_P(ResizeSchedulerTest, PropagateSliceToInputsWithReshape2) {
     const auto& heuristic_param =
         runtime->schedulerHeuristics()->heuristicsList().front();
     EXPECT_EQ(heuristic_param->scheduler_type, SchedulerType::Resize);
-    Fusion* scheduled_fusion =
-        runtime->executors().at(0)->as<KernelExecutor>()->kernel();
+    auto scheduled_fusion = runtime->executors()
+                                .at(0)
+                                ->as<KernelExecutor>()
+                                ->compiledKernel()
+                                ->kernel();
     checkLoopDomainEquivalence(
         scheduled_fusion->outputs().at(0)->as<TensorView>());
   }
@@ -4431,8 +4440,11 @@ TEST_P(ResizeSchedulerTest, PropagateMultipleSlicesToInputs1) {
     const auto& heuristic_param =
         runtime->schedulerHeuristics()->heuristicsList().front();
     EXPECT_EQ(heuristic_param->scheduler_type, SchedulerType::Resize);
-    Fusion* scheduled_fusion =
-        runtime->executors().at(0)->as<KernelExecutor>()->kernel();
+    auto scheduled_fusion = runtime->executors()
+                                .at(0)
+                                ->as<KernelExecutor>()
+                                ->compiledKernel()
+                                ->kernel();
     checkLoopDomainEquivalence(
         scheduled_fusion->outputs().at(0)->as<TensorView>());
   }
@@ -4637,8 +4649,11 @@ TEST_F(ResizeSchedulerTest, PropagateMultipleSlicesToInputs3) {
   const auto& heuristic_param =
       runtime->schedulerHeuristics()->heuristicsList().front();
   EXPECT_EQ(heuristic_param->scheduler_type, SchedulerType::Resize);
-  Fusion* scheduled_fusion =
-      runtime->executors().at(0)->as<KernelExecutor>()->kernel();
+  Fusion* scheduled_fusion = runtime->executors()
+                                 .at(0)
+                                 ->as<KernelExecutor>()
+                                 ->compiledKernel()
+                                 ->kernel();
   checkLoopDomainEquivalence(
       scheduled_fusion->outputs().at(0)->as<TensorView>());
 }
@@ -4777,8 +4792,11 @@ TEST_P(ResizeSchedulerTest, PropagateMultipleSlicesToInputs5) {
     const auto& heuristic_param =
         runtime->schedulerHeuristics()->heuristicsList().front();
     EXPECT_EQ(heuristic_param->scheduler_type, SchedulerType::Resize);
-    Fusion* scheduled_fusion =
-        runtime->executors().at(0)->as<KernelExecutor>()->kernel();
+    auto scheduled_fusion = runtime->executors()
+                                .at(0)
+                                ->as<KernelExecutor>()
+                                ->compiledKernel()
+                                ->kernel();
     checkLoopDomainEquivalence(
         scheduled_fusion->outputs().at(0)->as<TensorView>());
   }
@@ -4966,8 +4984,11 @@ TEST_P(ResizeSchedulerTest, SliceRotateCat) {
     const auto& heuristic_param =
         runtime->schedulerHeuristics()->heuristicsList().front();
     EXPECT_EQ(heuristic_param->scheduler_type, SchedulerType::Resize);
-    Fusion* scheduled_fusion =
-        runtime->executors().at(0)->as<KernelExecutor>()->kernel();
+    auto scheduled_fusion = runtime->executors()
+                                .at(0)
+                                ->as<KernelExecutor>()
+                                ->compiledKernel()
+                                ->kernel();
     checkLoopDomainEquivalence(
         scheduled_fusion->outputs().at(0)->as<TensorView>());
   }
@@ -5109,8 +5130,11 @@ TEST_P(ResizeSchedulerTest, SliceRotateCatResidual) {
     const auto& heuristic_param =
         runtime->schedulerHeuristics()->heuristicsList().front();
     EXPECT_EQ(heuristic_param->scheduler_type, SchedulerType::Resize);
-    Fusion* scheduled_fusion =
-        runtime->executors().at(0)->as<KernelExecutor>()->kernel();
+    auto scheduled_fusion = runtime->executors()
+                                .at(0)
+                                ->as<KernelExecutor>()
+                                ->compiledKernel()
+                                ->kernel();
     checkLoopDomainEquivalence(
         scheduled_fusion->outputs().at(0)->as<TensorView>());
   }
@@ -5296,8 +5320,11 @@ TEST_P(ResizeSchedulerTest, PropagatePadToInputs) {
     const auto& heuristic_param =
         runtime->schedulerHeuristics()->heuristicsList().front();
     EXPECT_EQ(heuristic_param->scheduler_type, SchedulerType::Resize);
-    Fusion* scheduled_fusion =
-        runtime->executors().at(0)->as<KernelExecutor>()->kernel();
+    auto scheduled_fusion = runtime->executors()
+                                .at(0)
+                                ->as<KernelExecutor>()
+                                ->compiledKernel()
+                                ->kernel();
     checkLoopDomainEquivalence(
         scheduled_fusion->outputs().at(0)->as<TensorView>());
   }
@@ -5397,8 +5424,11 @@ TEST_P(ResizeSchedulerTest, PropagateCatToInputs) {
     const auto& heuristic_param =
         runtime->schedulerHeuristics()->heuristicsList().front();
     EXPECT_EQ(heuristic_param->scheduler_type, SchedulerType::Resize);
-    Fusion* scheduled_fusion =
-        runtime->executors().at(0)->as<KernelExecutor>()->kernel();
+    auto scheduled_fusion = runtime->executors()
+                                .at(0)
+                                ->as<KernelExecutor>()
+                                ->compiledKernel()
+                                ->kernel();
     checkLoopDomainEquivalence(
         scheduled_fusion->outputs().at(0)->as<TensorView>());
   }
